feat(vibe-index): the wide-integer class boundary is domain honesty — counters ride as strings, counts as checked uint32 (B-095)
The rule «wider than 32 bits rides as a decimal string» met the
counter fields and the owner accepted the recommendation: the class
boundary is the DOMAIN, not the Rust type, recorded as an ADR-part
beside the rule itself (PROP-044 §4.2b, per the fresh B-007
convention). One question per field — does its domain honestly fit
uint32?

Three admin counters whose domains do not (uptime_seconds,
requests_total, mutations_total) now ride as canonical decimal
strings through the existing wire_decimal seam, their schemas say
string, and their «uint32 though u64» deviation notes die.
Twenty-two count-class fields across both envelope families keep
their uint32 schemas and earn them: wire_count::checked_u32 replaces
every silent `as` with a checked conversion that fails loudly on the
unreachable overflow — the CLI with a WireCountOverflow diagnostic,
the HTTP surface with 400/500 in the standard ErrorResponse — and
their softened deviation notes now truthfully say «bounded by
construction, checked at the boundary». The rider lands too: the
bare axum 404 fallback answers with the same ErrorResponse envelope
as every other error, pinned by an E2E test alongside one proving
offset=2^32 yields a 400 envelope, never a truncation.

Worker: codex (P2-FSYNC); gates: vibe-index 375 passed, parity and
golden suites green, clippy clean, codegen byte-idempotent, wire-diff
honestly reporting 11 schema + 1 corpus shifts under public=false.
The new boundary fact judged confirmed and sealed. B-095 closes.

+++ crates/vibe-index/src/index/search.rs
@@ -31,7 +31,7 @@ pub struct SearchHit {
     pub group: Group,
     pub name: String,
     pub latest_stable: Option<semver::Version>,
-    pub score: u32,
+    pub score: usize,
     pub matched_tokens: Vec<String>,
     pub description: Option<String>,
 }
@@ -99,7 +99,7 @@ pub fn search(index: &Index, query: &str, kind_filter: Option<PackageKind>) -> V
                 group: pkg.group.clone(),
                 name: pkg.name.clone(),
                 latest_stable: usable_latest_stable(pkg).cloned(),
-                score: matched.len() as u32,
+                score: matched.len(),
                 matched_tokens: matched.into_iter().cloned().collect(),
                 description: latest.description.clone(),
             },
